Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
clerk-sdk-python
email-validator==2.2.0
python-dotenv==1.0.1
python-multipart
